Add files via upload
This SQL file is totally conceptual. The idea was to make Shin Megami Tensei Digital Devil Saga into a tabletop like game similar to Dungeons and Dragons. The tables include characters, NPCs, items, locations, monsters, and quests. I chose this combo concept for my table because aside from being generally interested in them both, I believe having data laid out as such in these tables and queries can be very helpful for ‘players’ to find their way and create an adventure without the complication of researching each individual piece that is provided. The queries sort of a lot of frequently asked questions when it comes to this general game topic (player levels, npc languages, average levels, difficult enemies for a further challenge, etc). With these tables or at least in theory, players can sort and organize for a more concise playthrough of whatever this hypothetical game would be.